// backend/src/models/master-data.model.ts
export class ReqBasicInformation {
  namaWebsite: string;
  tagline: string;
  deskripsi: string;
}

export class ReqcontactService {
  telpon: string;
  email: string;
  instagram: string;
  websiteUtama: string;
}

export class MasterDataResponse {
  unggahBerkas: boolean;
  bukaPendaftaran: boolean;
  modePemeliharaan: boolean;
  namaWebsite: string;
  tagline: string;
  deskripsi: string;
}